Revert "[MM-38649] Put top bar in a BrowserView (#1758)"
This reverts commit c914903df918affff951c6279fb6308b6926df65.
