+++ src/main/menus/app.ts
@@ -142,7 +142,14 @@ function createTemplate(config: Config) {
             return 'Ctrl+Shift+I';
         })(),
         click(item: Electron.MenuItem, focusedWindow?: WebContents) {
-            WindowManager.openAppWrapperDevTools(focusedWindow);
+            if (focusedWindow) {
+                // toggledevtools opens it in the last known position, so sometimes it goes below the browserview
+                if (focusedWindow.isDevToolsOpened()) {
+                    focusedWindow.closeDevTools();
+                } else {
+                    focusedWindow.openDevTools({mode: 'detach'});
+                }
+            }
         },
     }, {
         label: 'Developer Tools for Current Tab',
